More user-friendly tidal dissipation notebook + attribution to shallow regions
I've reorganized the notebook around dictionaries for organizing experiment
metadata/datasets and now use doralite and gfdl_utils packages to more
adaptively find full data paths on PP/AN, rather than relying on hard-coded data
paths for every experiment and variable.

I've also added root-level `data/` and `figures/` directories to hold processed data
files that are essentially plot-ready. In this case, they are small enough to actually be
directly uploaded to the github repository, meaning that other users do not need to
recompute them to make the plots.

The notebook checks for the paths to these files. If the processed data files already exist
in the `data/` directory, then it simply reads them and goes straight to plotting. Otherwise,
it uses `gfdl_utils` to issue the necessary `dmget` commands to read them from `/archive/` and
do the necessary analysis to produce the processed data files.